refactor(Limiter): execute first job after next tick & refactor
This allows you to order your jobs, if needed, including `unshift()`
to the front of the queue, before the jobs start running.

This should align the limiter closer to programmer expectations, but
is technically breaking: the current code will immediately begin executing
the first job as soon as it is pushed.

This change also fixes a few minor bugs related to ordering & sync jobs:

- If an `onDone()` callback were added before any jobs were added in the same
tick, it would be immediately called.
- If a synchronous job were added, it would immediately execute completely
and potentially call `onDone()`.